Hello everyone,
Just checking if everyone is reading this.
We have to finish the release by this Wednesday and you are going
to make this happen! Thanks. Here is a list where we would be putting
who is doing what and what has to be done. If you run into additional
problems while doing your tasks just add the new ones to the list.



git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@405 02bb5f2a-90ca-11de-bb24-8aedcc835bc8